build: Drop release-please support for .github/workflows package (#51)
Release please has decided it really wants to write a changelog for this
package. But we know from a [previous open issue][1], this isn't
possible.

We only have this as a package as a way to try and version control the
stale and lint title workflows. But the burden this is now causing is
less than the benefit we receive from automating the release of two
workflows which nearly never change.

Going forward, if we want to make changes to those workflows, we can
manually tag them like the good ole days. It will remind us to stay
thankful for the automations we do have. And once the [the issue][1] is
fixed, we can restore support.

[1]: https://github.com/googleapis/release-please-action/issues/938